pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.1.3",
    "datasets>=3.3.2",
    "google>=3.0.0",
    "google-generativeai>=0.8.4",
    "mistralai>=1.5.1",
    "ollama>=0.4.7",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "pypdf2>=3.0.1",
    "requests>=2.32.3",
    "setuptools>=75.8.2",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.49.0",
]
